LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/animation-timeline-deferred-expected.txt


FAIL Animation.timeline returns attached timeline promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'animations[0].timeline')"
FAIL Animation.timeline returns null for inactive deferred timeline assert_equals: expected null but got object "[object ScrollTimeline]"
FAIL Animation.timeline returns null for inactive (overattached) deferred timeline assert_equals: expected null but got object "[object ScrollTimeline]"
